requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
CacheControl==0.12.11
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
discord.py @ git+https://github.com/Rapptz/discord.py@5456c0cae944f5e0dffb5de19cf18175e0d73011
firebase-admin==5.2.0
frozenlist==1.3.0
google-api-core==2.7.2
google-api-python-client==2.45.0
google-auth==2.6.5
google-auth-httplib2==0.1.0
google-cloud-core==2.3.0
google-cloud-firestore==2.4.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpcio==1.44.0
grpcio-status==1.44.0
httplib2==0.20.4
idna==3.3
msgpack==1.0.3
multidict==6.0.2
proto-plus==1.20.3
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.0
pyparsing==3.0.8
python-dotenv==0.20.0
requests==2.27.1
rsa==4.8
six==1.16.0
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
yarl==1.7.2
